types: creating quota unified type.
Now, both UserQuotaService and AppQuotaService uses the same type Quota. For now, both 'app' type and 'user' type will import Quota. Can be changed in the future, if needed to remove importing between types of different packages.
